pyproject.toml
[project]
name = "openai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "openai-agents>=0.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]
